Update index view with style changes and text additions.
For the left and right div elements, they need to be in their own
container, otherwise Gridism won't apply padding correctly and the
left div will not be aligned correctly. Also, the font styling was
modified slightly to add some spacing in between the letters of the
names, date, and location text.

- Modify layout of details div under card section.
- Remove spacing between city, state, names, and date.
- Wrap the details div in a grid (Gridism).
- Include new section, announcement, for some more information.
- Set background-color to white for body and remove padding.
- Add background-color to card section and add padding from body.
- Write out announcement text about upcoming things to expect.
- Set a box-shadow on the image selector within cover class.
- Add letter-spacing to the text in the content area.