Animate the index view with SVG and fade-in.
Our Save the Date design had a great looking
heart shape that I thought would look great
on the site. This adds the SVG image on page
load, animates it, then fades out to load the
actual content. It will only be animated on the
index view.

Note: This is a squashed commit from a feature branch
      called fade-in-svg.

- Draw the heart SVG then load the page.
- Hide all main content items initially.
- After slight delay, animate the SVG image.
- Add heart designed for Save the Date as SVG.
- Update readme's disclaimer to mention the designer.
- Create new partial, _heart, that stores the SVG in jade syntax.