Resolve issue introduced by fix around animation and nav.
This is a merge from the 'development' branch.

The previous fix just added a new element id for the
mobile nav. This resolved the issue in the short-term
but caused an issue with the menu-toggle icon not being
shown correctly.

The real issue was because the animation hides and shows
the main elements by the gridism class, unit. To fix the
issue, the best solution was to add a new class element
to each of the main elements that need to be hidden and
shown based on the animation.

- Add container class to main elements to hide/show.
- Update javascript with container class.