Adds an RSVP popup to the home page.
Still needs the proper text and some styling, but functionality is there.

Also, when a user RSVPs, we need to set a cookie so they don't see it again.

It might make more sense to put this functionality in a banner instead. Less intrusive, probably won't need any JS, etc.