Change data structure for Contact boolean values (#12).
To ensure it was easy to tell whether a contact had
entered RSVP information before and could select the
state of whether they are coming or not easily, the
schema needed to switch the isAttendingBigDay and
isAttendingRehearsalDinner data types to an integer.
They will use -1 if it has not been set (or null).

- Update all references to both properties to use integer.
- Change properties on Ember controller for contact.
- Disable the Next button and style differently.
- Handle state better based on what the user has/hasn't selected.